added presignedurl

+++ src/main/java/com/beeswork/balanceaccountservice/dto/account/PhotoDTO.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Setter
 public class PhotoDTO {
     private String key;
-    private Long sequence;
+    private int sequence;
 
 }
